"Heavy" refactoring and documenting
