+++ package.json
@@ -24,13 +24,13 @@
     "run-demo": "npm run prepare && python3 -m http.server --directory demo 12345"
   },
   "dependencies": {
-    "@babel/parser": "^7.23.9",
-    "@types/geojson": "7946.0.10"
+    "@babel/parser": "^7.23.9"
   },
   "peerDependencies": {
     "ol": "6 || 7 || 8"
   },
   "devDependencies": {
+    "@types/geojson": "7946.0.10",
     "@geoblocks/print": "^0.7.4",
     "@geoblocks/recast-utils": "^0.1.0",
     "@typescript-eslint/eslint-plugin": "6.20.0",
