+++ src/BaseCustomizer.ts
@@ -2,32 +2,79 @@ import type {Geometry} from 'ol/geom.js';
 import type {State} from 'ol/layer/Layer.js';
 import type {WMTS} from 'ol/source.js';
 import type {Image, Stroke} from 'ol/style.js';
-import type {
-  MapFishPrintSymbolizerLine,
-  MapFishPrintSymbolizerPoint,
-  MapFishPrintWmtsLayer,
-} from './mapfishprintTypes';
+import type {MFPSymbolizerLine, MFPSymbolizerPoint, MFPWmtsLayer} from './types';
+import type {Feature as GeoJSONFeature} from 'geojson';
 
+/**
+ * The customizer allows to customize some transformations.
+ * It also defines the print extent.
+ */
 export default class BaseCustomizer {
   readonly printExtent: number[];
 
+  /**
+   *
+   * @param printExtent The extent to print (useful for MVT / static image layers)
+   */
   constructor(printExtent: number[]) {
+    // FIXME: can not this be passed with the other options in createSpec?
     this.printExtent = printExtent;
   }
 
+  /**
+   *
+   * @param layerState
+   * @return true to convert this layer, false to skip it
+   */
   layerFilter(layerState: State): boolean {
     return true;
   }
 
+  /**
+   * Decide to skip some geometries.
+   * Useful to avoid sending features outside the print extend on the wire.
+   * @param geometry
+   * @return true to convert this feature, false to skip it
+   */
   geometryFilter(geometry: Geometry): boolean {
+    // FIXME: shouldn't we provide some reasonable defaults here?
+    // For ex:
+    // - define a buffer of X pixels and remove all points outside it;
+    // - only keep lines / polygons that intersect it
+    // Cf schm for some code.
     return true;
   }
 
-  feature(layerState: State, feature: GeoJSON.Feature) {}
+  /**
+   * Can be used to add / remove properties to features
+   * @param layerState
+   * @param feature converted feature
+   */
+  feature(layerState: State, feature: GeoJSONFeature) {}
 
-  line(layerState: State, symbolizer: MapFishPrintSymbolizerLine, stroke: Stroke) {}
+  /**
+   * Can be used to manipulate the line symbolizers
+   * @param layerState
+   * @param symbolizer
+   * @param stroke
+   */
+  line(layerState: State, symbolizer: MFPSymbolizerLine, stroke: Stroke) {}
 
-  point(layerState: State, symbolizer: MapFishPrintSymbolizerPoint, image: Image) {}
+  /**
+   * Can be used to manipulate the image symbolizers
+   * @param layerState
+   * @param symbolizer
+   * @param image
+   */
+  point(layerState: State, symbolizer: MFPSymbolizerPoint, image: Image) {}
 
-  wmtsLayer(layerState: State, wmtsLayer: MapFishPrintWmtsLayer, source: WMTS) {}
+  /**
+   * Can be used to manipulate a converted WMTS layer
+   * @param layerState
+   * @param wmtsLayer
+   * @param source
+   */
+  wmtsLayer(layerState: State, wmtsLayer: MFPWmtsLayer, source: WMTS) {}
+  // FIXME: does it really makes sense?
+  // Why isn't it done on an extended BaseEncoder instead?
 }
