+++ src/index.ts
@@ -1,6 +1,6 @@
-export * from './mapfishprintTypes';
+export * from './types';
 export {default as BaseCustomizer} from './BaseCustomizer';
-export * from './MapfishPrintBaseEncoder';
-export {default as MapfishPrintBaseEncoder} from './MapfishPrintBaseEncoder';
+export * from './MFPEncoder';
+export {default as MFPEncoder} from './MFPEncoder';
 export * from './VectorEncoder';
-export * from './mapfishprintUtils';
+export * from './utils';
